+++ deploy.sh
@@ -405,7 +405,11 @@ elif [ $1 = "linux" ]; then
     cp -r "$path"/vlc $deploy
 
     cp "$path"/libvlc*.so* $deploy
-    cp "$path"/libidn.so*  $deploy
+
+    if [ -f "$path"/libidn.so* ]; then
+
+        cp "$path"/libidn.so* $deploy
+    fi
 fi
 
 echo "-------------"
